consumer/requirements.txt
kafka-python
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-kafka-python
opentelemetry-exporter-jaeger
faker
